+++ ShellKrypt.Core/CryptoTools/IPasswordStrengthService.cs
@@ -1,0 +1,6 @@
+namespace ShellKrypt.Core.CryptoTools;
+
+public interface IPasswordStrengthService
+{
+    PasswordStrengthAssessment AssessPasswordStrength(string? password);
+}